Add clearFunction, switch statement to operator
Add function to clear current numbers displayed and add
switch statement to the operator function where the operator
clicked is evaluated. This is to see if equals has been pressed or not.
Add code to see if a operator has been chosen in the selectorFunction

Change button id to class on the operator buttons

Do minor tweaks to the CSS to start getting the layout and style